Move AlarmSeen to new auxiliarymodel package
